修改记录2017.09.28   #{      @Time 09:20  第一：添加六大現場采样页面展示效果和上传效果。第二：添加时间选择器。第三：添加完成5秒弹窗。第四：修改GIS地图展示中心位置   #}

+++ Sampling/app/src/main/java/com/xiangxun/sampling/ui/presenter/SamplingDBPresenter.java
@@ -41,45 +41,71 @@ public class SamplingDBPresenter {
     }
 
 
-    public void upAll(final String id, final String pointId) {
-        if (TextUtils.isEmpty(id)) {
+    public void upAll(final List<String> ids, final boolean isAll) {
+        if (ids == null){
             ToastApp.showToast("上传id不能为空");
             return;
         }
-        final List<SenceSamplingSugar> sence = SugarRecord.find(SenceSamplingSugar.class, "samplingId = ?", id);
-        final List<MediaSugar> sugar = SugarRecord.find(MediaSugar.class, "samplingId = ?", id);
+        if (ids.size() == 0){
+            ToastApp.showToast("上传id不能为空");
+            return;
+        }
         //构建body
         final MultipartBody.Builder builder = new MultipartBody.Builder().setType(MultipartBody.FORM);
-        builder.addFormDataPart("", id, RequestBody.create(MediaType.parse("image/*"), ""));
-        if (sugar != null && sugar.size() != 0) {
-            for (MediaSugar media : sugar) {
-                File file = new File(media.getUrl());
-                if ("image".equals(media.getType())) {
-                    builder.addFormDataPart(media.getType(), id + "@" + file.getName(), RequestBody.create(MediaType.parse("image/*"), file));
-                }
-                if ("video".equals(media.getType())) {
-                    builder.addFormDataPart(media.getType(), id + "@" + file.getName(), RequestBody.create(MediaType.parse("video/*"), file));
-                }
+        builder.addFormDataPart("", "", RequestBody.create(MediaType.parse("image/*"), ""));
+        for (String id:ids) {
+            if (TextUtils.isEmpty(id)){
+                //假如中间一个id为空，则跳出继续下个id查找。
+                continue;
             }
-        } else {
-            ToastApp.showToast("上传图片和视频不能为空");
-            return;
+            List<MediaSugar> sugar = SugarRecord.find(MediaSugar.class, "samplingId = ?", id);
+            if (sugar != null && sugar.size() != 0) {
+                for (MediaSugar media : sugar) {
+                    File file = new File(media.getUrl());
+                    if ("image".equals(media.getType())) {
+                        builder.addFormDataPart(media.getType(), id + "@" + media.getSamplingCode() + "@" + file.getName(), RequestBody.create(MediaType.parse("image/*"), file));
+                    }
+                    if ("video".equals(media.getType())) {
+                        builder.addFormDataPart(media.getType(), id + "@" + media.getSamplingCode() + "@" + file.getName(), RequestBody.create(MediaType.parse("video/*"), file));
+                    }
+                }
+            } else {
+                //这个采样记录没有图片和视频信息。
+                continue;
+            }
         }
         RequestBody body = builder.build();
         biz.onStart(loading);
         biz.upAll(body, new FrameListener<ResultPointData>() {
             @Override
             public void onSucces(ResultPointData result) {
-                biz.onStop(loading);
                 //上传成功后，删除本地数据库信息
-                for (MediaSugar media : sugar) {
-                    media.delete();
+                for (String id:ids) {
+                    if (TextUtils.isEmpty(id)){
+                        //假如中间一个id为空，则跳出继续下个id查找。
+                        continue;
+                    }
+                    List<MediaSugar> sugar = SugarRecord.find(MediaSugar.class, "samplingId = ?", id);
+                    //上传成功后，删除图片数据库里面的信息。
+                    if (sugar != null && sugar.size() != 0) {
+                        for (MediaSugar media : sugar) {
+                            media.delete();
+                        }
+                    }
+                    List<SenceSamplingSugar> sence = SugarRecord.find(SenceSamplingSugar.class, "samplingId = ?", id);
+                    if (sence != null && sence.size() != 0) {
+                        for (SenceSamplingSugar media : sence) {
+                            media.delete();
+                        }
+                    }
                 }
-                for (SenceSamplingSugar media : sence) {
-                    media.delete();
+                if (isAll){
+                    //删除文件信息
+                    RecursionDeleteFile(new File(Api.VIDEO));
                 }
-                RecursionDeleteFile(new File(Api.SENCE.concat(pointId)));
+                biz.onStop(loading);
                 view.onUpSuccess();
+
             }
 
             @Override
